Fix layout in Internet Explorer 7/8
IE prior to version 9 will create an empty inline element in the
DOM for both the opening and closing HTML5 elements. The recommended
way to workaround this is use div tags to wrap the semantic
elements and use those for styling.

Whilst [Tantek Çelik’s][1] [preferred method][2] is to wrap the HTML5
element around the styling div tag, I’ve gone for wrapping the
div tag around the semantic element as this means the metrics of
both elements match in HTML5 aware browsers.

Another option was to use the [HTML5 shiv][3] JS library but I
rejected this as being unnecessary for simple pages such as these.

[1]: http://tantek.com/
[2]: http://tantek.com/presentations/2010/11/html5-now/
[3]: http://code.google.com/p/html5shim/