Replace image headers with styled HTML
This is not currently a complete replacement as a new font to replace
Interstate in the graphic headers still needs to be found. Once a replacement
font is found the stylesheet will need to be tweaked to be more visually
pleasing. For example, there are some line wraps that should be prevented.

This sitemap changes were also committed here, although I'm not sure if this is
correct since there were no content changes to those pages.

Combine duplicate header styles