Uniform the layout of /security with the site
Every page of the website has the same layout (header/body) except
for the Security page. This commit fixes that, styling the body as the rest of
the site and adding the following header:

>Security policy
>
>Did you find a security vulnerability in Ruby on Rails?
>Follow these steps to safely report the issue to the core team.
>You might even be eligible for a bounty.

The only content added is the header; let me know if you agree.
Please be aware that the Security page is not linked anywhere from the rest
of the site.

It is linked only from [this paragraph in the guides](http://guides.rubyonrails.org/contributing_to_ruby_on_rails.html#special-treatment-for-security-issues),
so the header should somehow following that paragraph which says:

> 1.3 Special Treatment for Security Issues
> Please do not report security vulnerabilities with public GitHub issue reports. The Rails security policy page details the procedure to follow for security issues.