Don’t have styles specific to the Quotes page
Removing the `class="quotes"` in the quotes page is the last step
to guarantee that all the pages have the same layout (which can
therefore be extracted into the "default" layout).

The `class="quotes"` was only there to style `<blockquote`>, but
their style can be defined without being specific to that page.