Load Bootstrap CSS, keep the current layout intact
The first step for a responsive design is to use a CSS framework that
supports grid (or to create one). I suggest we use the most widely used
at the time: Bootstrap CSS.

This commit loads Bootstrap CSS while keeping the current layout intact.

The following commits will gradually replace the element of the current
layout (such as `<div role="main"><section>`) with responsive elements
provided by Bootstrap (such as `<div class="container">`).

I'm open to discussion about whether it’s okay load Bootstrap CSS from
a CDN or include the file in the repository.

I'm in favor of using a CDN, because it guarantees faster loading time +
the file might already be in the browser’s cache + does not require the
CSS from the same host (GitHub Pages) that serves the HTML files.