Responsive design: navigation menu
This is the first commit to gradually make the web site responsive (#46).
This commit *only* deals with the navigation menu at the top, and uses
Bootstrap grid system so that it:

* looks the same in a normal-sizes browser window
* displays a collapsed dropdown menu in small-sized windows

Interesting points of discussion about this commit:

1) the collapsed dropdown requires bootstrap.js -- as in the previous commit, up for discussion whether using a CDN is ok
2) bootstrap.js requires a more modern version of jquery, so 1.11.1 is used instead of 1.7.1
3) some CSS rules have been added to transform the "bootstrap-styled" navbar into the same style currently used by the site

Regarding the last point, I just want to iterate that this commit intends not
to change the current style/layout at all, just to make it responsive.
I'm all for removing some of those extra CSS rules if we want to change the
design, but it's outside of the scope of the current PR.

Attachments show how the site and the navigation looks like before/after the commit.