Don’t use `box-sizing: content-box` in CSS
In order to use Bootstrap grid system, the box-sizing should be the
default, that is, any measure of width or height should be inclusive
of padding and border (but not margin).

This commit replaces the box-sizing value without affecting the web
site, by updating all the width/height values to include paddings and
borders.